+++ lib/ns-capabilities.ts
@@ -3,6 +3,7 @@ import { INsCapabilities, AutomationName } from "./interfaces/ns-capabilities";
 import { resolveCapabilities } from "./capabilities-helper";
 import { getAppPath, fileExists, logErr, logInfo, logError } from "./utils";
 import { IDevice } from "mobile-devices-controller";
+import { IDeviceManager } from "./interfaces/device-manager";
 
 export class NsCapabilities implements INsCapabilities {
     private _projectDir;
@@ -19,8 +20,6 @@ export class NsCapabilities implements INsCapabilities {
     private _reuseDevice;
     private _devMode;
     private _runType;
-    private _isAndroid;
-    private _isIOS;
     private _isSauceLab;
     private _appName: string;
     private _appPath: string;
@@ -35,7 +34,8 @@ export class NsCapabilities implements INsCapabilities {
     private _attachToDebug: boolean;
     private _startSession: boolean;
     private _sessionId: string;
-    private exceptions: Array<string> = new Array();
+    private _deviceManager: IDeviceManager;
+    private _exceptions: Array<string> = new Array();
 
     constructor() {
         this._projectDir = parser.projectDir;
@@ -46,29 +46,31 @@ export class NsCapabilities implements INsCapabilities {
         this._port = parser.port;
         this._verbose = parser.verbose;
         this._appiumCapsLocation = parser.appiumCapsLocation;
-        this._appiumCaps = resolveCapabilities(this._appiumCapsLocation, parser.runType, parser.projectDir);
+        this._relaxedSecurity = parser.relaxedSecurity;
+        this._cleanApp = parser.cleanApp;
+        this._attachToDebug = parser.attachToDebug;
+        this._sessionId = parser.sessionId;
+        this._startSession = parser.startSession;
+        if (!this._attachToDebug && !this._sessionId) {
+            this._appiumCaps = resolveCapabilities(this._appiumCapsLocation, parser.runType, parser.projectDir);
+        }
         this._testFolder = parser.testFolder;
         this._storage = parser.storage;
         this._testReports = parser.testReports;
         this._reuseDevice = parser.reuseDevice;
         this._devMode = parser.devMode;
         this._runType = parser.runType;
-        this._isAndroid = this.isAndroidPlatform();
-        this._isIOS = !this._isAndroid;
         this._isSauceLab = parser.isSauceLab;
         this._ignoreDeviceController = parser.ignoreDeviceController;
         this._wdaLocalPort = parser.wdaLocalPort;
         this._path = parser.path;
-        this._relaxedSecurity = parser.relaxedSecurity;
-        this._cleanApp = parser.cleanApp;
-        this._attachToDebug = parser.attachToDebug;
-        this._sessionId = parser.sessionId;
-        this._startSession = parser.startSession;
-        this.setAutomationName();
-        this.resolveApplication();
-        this.checkMandatoryCapabiliies();
-        this.throwExceptions();
-        this.shouldSetFullResetOption();
+        if (!this._attachToDebug && !this._sessionId) {
+            this.setAutomationName();
+            this.resolveApplication();
+            this.checkMandatoryCapabiliies();
+            this.throwExceptions();
+            this.shouldSetFullResetOption();
+        }
     }
 
     get path() { return this._path; }
@@ -80,14 +82,15 @@ export class NsCapabilities implements INsCapabilities {
     get verbose() { return this._verbose; }
     get appiumCapsLocation() { return this._appiumCapsLocation; }
     get appiumCaps() { return this._appiumCaps; }
+    set appiumCaps(appiumCaps) { this._appiumCaps = appiumCaps; }
     get testFolder() { return this._testFolder; }
     get storage() { return this._storage; }
     get testReports() { return this._testReports; }
     get reuseDevice() { return this._reuseDevice; }
     get devMode() { return this._devMode; }
     get runType() { return this._runType; }
-    get isAndroid() { return this._isAndroid; }
-    get isIOS() { return this._isIOS; }
+    get isAndroid() { return this.isAndroidPlatform(); }
+    get isIOS() { return !this.isAndroid; }
     get isSauceLab() { return this._isSauceLab; }
     get automationName() { return this._automationName; }
     get appPath() { return this._appPath; }
@@ -105,6 +108,8 @@ export class NsCapabilities implements INsCapabilities {
     get sessionId() { return this._sessionId; }
     set sessionId(sessionId: string) { this._sessionId = sessionId; }
     get startSession() { return this._startSession; }
+    get deviceManager() { return this._deviceManager; }
+    set deviceManager(deviceManager: IDeviceManager) { this._deviceManager = deviceManager; }
 
     private isAndroidPlatform() { return this._appiumCaps.platformName.toLowerCase().includes("android"); }
 
@@ -133,7 +138,7 @@ export class NsCapabilities implements INsCapabilities {
                     this._automationName = AutomationName.XCUITest; break;
             }
         } else {
-            if (this._isAndroid) {
+            if (this.isAndroid) {
                 if (this.tryGetAndroidApiLevel() > 6 || (this.appiumCaps["apiLevel"] && this.appiumCaps["apiLevel"].toLowerCase().includes("p"))) {
                     this._automationName = AutomationName.UiAutomator2;
                 }
@@ -173,15 +178,15 @@ export class NsCapabilities implements INsCapabilities {
 
     private checkMandatoryCapabiliies() {
         if (!this.isSauceLab && !fileExists(this._appiumCaps.app)) {
-            this.exceptions.push("The application folder doesn't exist!");
+            this._exceptions.push("The application folder doesn't exist!");
         }
 
         if (!this._runType) {
-            this.exceptions.push("Missing runType! Please select one from appium capabilities file!");
+            this._exceptions.push("Missing runType! Please select one from appium capabilities file!");
         }
 
         if (!this._appiumCaps.platformName) {
-            this.exceptions.push("Platform name is missing! Please, check appium capabilities file!");
+            this._exceptions.push("Platform name is missing! Please, check appium capabilities file!");
         }
 
         if (!this._appiumCaps.platformVersion) {
@@ -189,17 +194,17 @@ export class NsCapabilities implements INsCapabilities {
         }
 
         if (!this._appiumCaps.deviceName && !this._appiumCaps.udid) {
-            this.exceptions.push("The device name or udid are missing! Please, check appium capabilities file!");
+            this._exceptions.push("The device name or udid are missing! Please, check appium capabilities file!");
         }
     }
 
     private throwExceptions() {
-        this.exceptions.forEach(msg => {
+        this._exceptions.forEach(msg => {
             logError(msg);
         });
 
-        if (this.exceptions.length > 0) {
-           process.exit(1);
+        if (this._exceptions.length > 0) {
+            process.exit(1);
         }
     }
 }