+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="4.0.1"></a>
+# [4.0.1]() (2018-07-20)
+
+### Features
+* **UIElement:** intorduce wd property
+* **AppiumDriver:** intorduce touchAction() 
+* **ElementHelper:** intorduce getXPathByTextAtributes() - returns element/s specific type or '//*' that meets all atrbutes for ios ["label", "value", "hint"] and for android ["content-desc", "resource-id", "text"]
+* --attachToDebug option requires only --port option
+* --sessionId options requires only --port oprtion
+
+
 <a name="4.0.0"></a>
 # [4.0.0]() (2018-07-02)
 
@@ -16,7 +27,7 @@
 * deviceLog ([1279323](https://github.com/NativeScript/nativescript-dev-appium/commit/1279323)). exposed driver.getDeviceLog() which will return device log and exposed driver.logDeviceLog(fileName, logtype) which will create such file with postfix `${fileName}_${logtype}.log` in reports folder under appName/deviceName
 * **android:** provide option "--relaxedSecurity" to enable relaxed security and execute shell commands using appium android driver (#126) ([ce780bf](https://github.com/NativeScript/nativescript-dev-appium/commit/ce780bf)), closes [#126](https://github.com/NativeScript/nativescript-dev-appium/issues/126)
 
-+### BREAKING CHANGES
+### BREAKING CHANGES
 
 * **android:** change default automation name if not specified in appium config  (#122) ([3ba0a1c](https://github.com/NativeScript/nativescript-dev-appium/commit/3ba0a1c)), closes [#122](https://github.com/NativeScript/nativescript-dev-appium/issues/122).